+++ src/main/java/com/ghgande/j2mod/modbus/net/TCPMasterConnection.java
@@ -16,8 +16,8 @@
 package com.ghgande.j2mod.modbus.net;
 
 import com.ghgande.j2mod.modbus.Modbus;
+import com.ghgande.j2mod.modbus.io.AbstractModbusTransport;
 import com.ghgande.j2mod.modbus.io.ModbusTCPTransport;
-import com.ghgande.j2mod.modbus.io.ModbusTransport;
 import com.ghgande.j2mod.modbus.util.ModbusLogger;
 
 import java.io.IOException;
@@ -174,7 +174,7 @@ public class TCPMasterConnection {
      *
      * @return the connection's <tt>ModbusTransport</tt>.
      */
-    public ModbusTransport getModbusTransport() {
+    public AbstractModbusTransport getModbusTransport() {
         return transport;
     }