+++ src/main/java/com/ghgande/j2mod/modbus/net/UDPMasterConnection.java
@@ -16,7 +16,7 @@
 package com.ghgande.j2mod.modbus.net;
 
 import com.ghgande.j2mod.modbus.Modbus;
-import com.ghgande.j2mod.modbus.io.ModbusTransport;
+import com.ghgande.j2mod.modbus.io.AbstractModbusTransport;
 import com.ghgande.j2mod.modbus.util.ModbusLogger;
 
 import java.net.InetAddress;
@@ -86,7 +86,7 @@ public class UDPMasterConnection {
      *
      * @return the connection's <tt>ModbusTransport</tt>.
      */
-    public ModbusTransport getModbusTransport() {
+    public AbstractModbusTransport getModbusTransport() {
         return terminal == null ? null : terminal.getTransport();
     }