+++ src/test/java/com/ghgande/j2mod/modbus/cmd/ReadFileRecordTest.java
@@ -18,9 +18,9 @@ package com.ghgande.j2mod.modbus.cmd;
 import com.ghgande.j2mod.modbus.ModbusException;
 import com.ghgande.j2mod.modbus.ModbusIOException;
 import com.ghgande.j2mod.modbus.ModbusSlaveException;
+import com.ghgande.j2mod.modbus.io.AbstractModbusTransport;
 import com.ghgande.j2mod.modbus.io.ModbusSerialTransport;
 import com.ghgande.j2mod.modbus.io.ModbusTransaction;
-import com.ghgande.j2mod.modbus.io.ModbusTransport;
 import com.ghgande.j2mod.modbus.msg.*;
 import com.ghgande.j2mod.modbus.msg.ReadFileRecordRequest.RecordRequest;
 import com.ghgande.j2mod.modbus.msg.ReadFileRecordResponse.RecordResponse;
@@ -51,7 +51,7 @@ public class ReadFileRecordTest {
     }
 
     public static void main(String[] args) {
-        ModbusTransport transport = null;
+        AbstractModbusTransport transport = null;
         ReadFileRecordRequest request;
         ReadFileRecordResponse response;
         ModbusTransaction trans;
@@ -69,7 +69,7 @@ public class ReadFileRecordTest {
         try {
             transport = ModbusMasterFactory.createModbusMaster(args[0]);
             if (transport instanceof ModbusSerialTransport) {
-                ((ModbusSerialTransport)transport).setReceiveTimeout(500);
+                transport.setTimeout(500);
                 if (System.getProperty("com.ghgande.j2mod.modbus.baud") != null) {
                     ((ModbusSerialTransport)transport).setBaudRate(Integer.parseInt(System.getProperty("com.ghgande.j2mod.modbus.baud")));
                 }