+++ src/test/java/com/ghgande/j2mod/modbus/cmd/WriteFileRecordTest.java
@@ -18,9 +18,9 @@ package com.ghgande.j2mod.modbus.cmd;
 import com.ghgande.j2mod.modbus.ModbusException;
 import com.ghgande.j2mod.modbus.ModbusIOException;
 import com.ghgande.j2mod.modbus.ModbusSlaveException;
+import com.ghgande.j2mod.modbus.io.AbstractModbusTransport;
 import com.ghgande.j2mod.modbus.io.ModbusSerialTransport;
 import com.ghgande.j2mod.modbus.io.ModbusTransaction;
-import com.ghgande.j2mod.modbus.io.ModbusTransport;
 import com.ghgande.j2mod.modbus.msg.ExceptionResponse;
 import com.ghgande.j2mod.modbus.msg.ModbusResponse;
 import com.ghgande.j2mod.modbus.msg.WriteFileRecordRequest;
@@ -55,7 +55,7 @@ public class WriteFileRecordTest {
     }
 
     public static void main(String[] args) {
-        ModbusTransport transport = null;
+        AbstractModbusTransport transport = null;
         WriteFileRecordRequest request;
         WriteFileRecordResponse response;
         ModbusTransaction trans;
@@ -74,7 +74,7 @@ public class WriteFileRecordTest {
         try {
             transport = ModbusMasterFactory.createModbusMaster(args[0]);
             if (transport instanceof ModbusSerialTransport) {
-                ((ModbusSerialTransport)transport).setReceiveTimeout(500);
+                transport.setTimeout(500);
                 ((ModbusSerialTransport)transport).setBaudRate(19200);
                 isSerial = true;